Behavioural-model/version_0.3/requirements.txt
pandas
numpy
scikit-learn
lightgbm
shap
pyyaml
joblib
matplotlib
